Adding support for Active Messages in OpenSHMEM (request-reply model)
A PE can initiate a call to an active message on a
remote PE using shmem_am_request. The request handler
can in turn choose to launch a reply handler on the
initial PE using shmem_am_reply. If this function is
not called by the request handler, the underlying SHMEM
implementation is responsible for launching a NULL
reply handler on its behalf (after the execution of the
handler body).